Refactoring: Separate out workflow related functionality
No semantic changes.
